Add files via upload
This is an example for Pearson's Correlation & Spearman's Correlation. While the Pearson's Correlation Coefficient quantifies the relationship between two variables, it is not able to capture a non linear relationship. Ths Spearman's Correlation Coefficient has an advantage that it is not impacted by non-linear variations as it converts all values to an equivalent rank. Thus the Spearman's Correlation Coefficient expresses better insight into the relationship between two variables.